hopeps_bank_email_backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
python-decouple==3.6
python-http-client==3.3.7
pytz==2022.6
requests==2.28.1
sendgrid==6.9.7
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
urllib3==1.26.13
